implemented transaction processing added a separate table for dispute resolutions. inserting resolutions for previously disputed transactions now fails moved transaction_processing from the executable file (the one in /bin) to its own module
