+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod db;
 pub mod errors;
 pub mod model;
+pub mod transaction_processor;
